Divide buildpack in two phases: triggering build and busy loop waiting for it
The idea is that, we'll add this buildpack twice in the list of buildpacks.
It will run before all other buildpacks, and as last one as well. So when
the first buildpack execution happens, if the manifests archive is not found,
we trigger a build on Harness and allow the other buildpacks to proceed.

The second time this buildpack runs, it will enter a busy loop waiting for the
manifests file. This allows us to reduce the amount of time we spend idle, waiting
in a busy loop.